requirements.txt
beautifulsoup4
pillow
torch
torchvision
flask